requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.2.post1
langchain-pinecone==0.0.1
pinecone-client==3.0.1
openai==1.8.0
pydantic==2.5.3
typing-extensions==4.9.0
uuid==1.30